Set up NDBC API proxy route.
The NDBC server doesn't seem to be working with CORS, at least not
on my Chrome.
Also, the data is an RSS feed (XML), so there are a few hoops to jump
through to convert it to JSON and send it to the front end.
There is some markup () in the description text.

So we'll use an express route to get the data, and clean it up into a
tidy JSON object to send to the browser.